backend/fixtures/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
alembic==1.12.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
redis==4.6.0
httpx==0.25.0
email-validator==2.0.0
faker==37.3.0
